user

+++ src/trpcs/routes/orders.route.ts
@@ -12,7 +12,9 @@ export const ordersRouter = router({
             start: z.string(),
             end: z.string()
         }))
-        .query(async ({ input: { start, end, status } }) => {
+        .query(async ({ input: { start, end, status }, ctx }) => {
+
+            console.log(ctx)
 
             const startDate = startOfDay(parse(start, 'yyyy-MM-dd', new Date()))
 
@@ -26,7 +28,8 @@ export const ordersRouter = router({
                     createdAt: {
                         gte: startDate,
                         lte: endDate,
-                    }
+                    },
+                    userid: ctx.userId
                 },
                 include: {
                     custumer: true, orderServices: {
@@ -40,7 +43,8 @@ export const ordersRouter = router({
         .query(async (opts) => {
             return await db.order.findFirst({
                 where: {
-                    id: opts.input
+                    id: opts.input,
+                    userid: opts.ctx.userId
                 },
                 include: {
                     custumer: true,
@@ -63,14 +67,14 @@ export const ordersRouter = router({
         }),
     delete: protectedProcedure
         .input(z.string())
-        .mutation(async ({ input }) => {
+        .mutation(async ({ input, ctx }) => {
 
             return await db.$transaction(async trx => {
-                await trx.orderProduct.deleteMany({ where: { orderid: input } });
-                await trx.orderProduct.deleteMany({ where: { orderid: input } });
-                await trx.orderPayment.deleteMany({ where: { orderId: input } });
+                await trx.orderService.deleteMany({ where: { orderId: input, userid: ctx.userId } });
+                await trx.orderProduct.deleteMany({ where: { orderid: input, userid: ctx.userId } });
+                await trx.orderPayment.deleteMany({ where: { orderId: input, userid: ctx.userId } });
 
-                return await db.order.delete({ where: { id: input } })
+                return await trx.order.delete({ where: { id: input, userid: ctx.userId } })
             })
 
         }),
@@ -83,6 +87,7 @@ export const ordersRouter = router({
                 data: {
                     total: 0,
                     custumerId: opts.input.custumerId,
+                    userid: opts.ctx.userId
                 }
             })
 
@@ -99,7 +104,7 @@ export const ordersRouter = router({
         .mutation(async opts => {
 
             const order = await db.order.findFirst({
-                where: { id: opts.input.id },
+                where: { id: opts.input.id, userid: opts.ctx.userId },
                 include: { orderServices: true, orderProducts: true }
             })
 
@@ -116,12 +121,13 @@ export const ordersRouter = router({
             }
 
             return await db.order.update({
-                where: { id: opts.input.id },
+                where: { id: opts.input.id, userid: opts.ctx.userId },
                 data: {
                     done: true, total, orderPayments: {
                         createMany: {
                             data: opts.input.payment.map(item => {
                                 return {
+                                    userid: opts.ctx.userId,
                                     methodId: item.methodId,
                                     value: item.value,
                                 }
@@ -141,34 +147,37 @@ export const ordersRouter = router({
         price: z.number(),
         description: z.string(),
     }))
-        .mutation(async ({ input }) => {
+        .mutation(async ({ input, ctx }) => {
             const order = await db.order.findFirst({ where: { id: input.orderId } })
 
             return db.orderService.create({
                 data: {
                     ...input,
-                    custumerId: order?.custumerId
+                    custumerId: order?.custumerId,
+                    userid: ctx.userId
                 }
             })
         }),
-    deleteService: protectedProcedure.input(z.string()).mutation(async ({ input }) => {
+    deleteService: protectedProcedure.input(z.string()).mutation(async ({ input, ctx }) => {
         return await db.orderService.delete({
             where: {
-                id: input
+                id: input,
+                userid: ctx.userId
             }
         })
     }),
-    deleteProduct: protectedProcedure.input(z.string()).mutation(async ({ input }) => {
+    deleteProduct: protectedProcedure.input(z.string()).mutation(async ({ input, ctx }) => {
         return await db.orderProduct.delete({
             where: {
-                id: input
+                id: input,
+                userid: ctx.userId
             }
         })
     }),
     addProduct: protectedProcedure.input(z.object({ orderId: z.string(), productId: z.string(), qty: z.number() }))
-        .mutation(async ({ input }) => {
-            const order = await db.order.findFirst({ where: { id: input.orderId }, include: { orderProducts: true } })
-            const product = await db.product.findFirst({ where: { id: input.productId } })
+        .mutation(async ({ input, ctx }) => {
+            const order = await db.order.findFirst({ where: { id: input.orderId, userid: ctx.userId }, include: { orderProducts: true } })
+            const product = await db.product.findFirst({ where: { id: input.productId, userid: ctx.userId } })
 
             if (!product || !order) return null;
 
@@ -177,7 +186,8 @@ export const ordersRouter = router({
             if (prevProduct) {
                 return await db.orderProduct.update({
                     where: {
-                        id: prevProduct.id
+                        id: prevProduct.id,
+                        userid: ctx.userId
                     },
                     data: {
                         qty: prevProduct.qty + input.qty,
@@ -192,7 +202,8 @@ export const ordersRouter = router({
                     productid: input.productId,
                     qty: input.qty,
                     price: product.price,
-                    total: input.qty * product.price
+                    total: input.qty * product.price,
+                    userid: ctx.userId
                 }
             })
         })
